Create PushDataFromExcel_UrbanskiModule
This macro file contains Tomasz Urbanski's code adapted to my company's local needs and paths and file names. 
It assigns text selected in Word, usually the PESEL number and finds it in "X:\Roesler\Excel\FW 1\custdb.xlsm" database file.
If the value is found, in other words, if the customer data are found in company's database, then his name and ID's number are pasted into the corresponding content control fields in the Word document.